refactor: separate spacing and layout tokens from color system
Split design tokens into focused, single-responsibility modules:

- colors.css: Only color palettes (primary, accent, success, warning, neutral)
- spacing.css: 4px-based spacing scale (--space-1 through --space-20)
- layout.css: Border radius and shadows for elevation system

Benefits:
- Better separation of concerns
- Clearer file organization
- Easier to maintain and extend
- Each system has its own test file and documentation

Files created:
- src/styles/spacing.css and spacing.test.ts
- src/styles/layout.css and layout.test.ts
- docs/spacing-system.md
- docs/layout-system.md

Files updated:
- src/styles/colors.css: Removed spacing, border radius, shadows
- src/styles/colors.test.ts: Removed non-color tests
- src/index.css: Import spacing.css and layout.css
- docs/color-system.md: Updated to focus only on colors

All tests passing (13 tests across colors, spacing, layout systems)
